moved specific endpoints to another file
